GUI logic for the controll of the GUI
Some new classes appeared and some of them will be removed in the nearest future, as they are placeholders for now.
The TurnController will handle the appearance of GUI during the turns. The Engine will control the flow of messages from the Game Core and into it.
Now I will need to get rid of the placeholders, refactor a bit and develop the TurnController further so it would actually affect the GUI.